requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
aiosqlite
databases
sqlalchemy